+++ ask-smapi-sdk/package.json
@@ -19,7 +19,6 @@
     "ask-smapi-model": "^1.0.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.1.2",
     "@types/mocha": "^5.0.0",
     "@types/node": "^9.6.1",
     "@types/sinon": "^4.3.0",
