+++ ask-smapi-sdk/tslint.json
@@ -160,13 +160,9 @@
     "typedef": [
       true,
       "call-signature",
-      //"arrow-call-signature",
       "parameter",
-      //"arrow-parameter",
       "property-declaration",
       "member-variable-declaration"
-      //"object-destructuring",
-      //"array-destructuring"
     ],
     "typedef-whitespace": [
       true,
